define and use some xacml utilities
the utilities deal mostly with the possibility to use
a shared_ptr for xacml entities with a deleter that allows
to release the raw pointer. In this way it is possible to
use the shared_ptr with a deleter that cleans up in case of a
failure (e.g. an exception). At the same time however it's
possible to release the raw pointer from the shared_ptr to pass
it to the xacml api, which will then own it.