use xacml utils also for the request
the xacml_request_t is created using the xacml utilities,
in particular the special deleter. The raw pointer for this
request, once initialized successfully, is passed to the
pep api, which replace the pointer with another one. This
returned pointer is managed with a shared ptr and the normal
deleter (xacml_request_delete).