Convert storm_util to C++
Fix https://issues.infn.it/jira/browse/STOR-1309
The change has caused get_supported_protocols to be called twice instead of once
during startup. The result could be cached in some way, but not as it was done
before with a global var. Note that the call is then done for each
GetTransferProtocols request. Can this also benefit from the caching or
can the list change at runtime?